requirements.txt
aiocontextvars==0.2.2
aiofiles==0.7.0
asgiref==3.4.1
async-exit-stack==1.0.1
async-generator==1.10
click==8.0.1
colorama==0.4.4
contextvars==2.4
dataclasses==0.8
fastapi==0.66.0
greenlet==1.1.0
h11==0.12.0
immutables==0.15
importlib-metadata==4.6.1
loguru==0.5.3
motor==2.4.0
passlib==1.7.4
pydantic==1.8.2
PyJWT==2.1.0
pymongo==3.11.4
SQLAlchemy==1.4.20
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
Werkzeug==2.0.1
win32-setctime==1.0.3
zipp==3.5.0
bcrypt==3.1.7